requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
python-decouple==3.8
whitenoise==6.6.0
gunicorn==21.2.0
reportlab==4.0.4
Pillow==10.0.1
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0
channels==4.0.0
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
django-notifications-hq==1.8.3
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
django-extensions==3.2.3
django-cors-headers==4.3.1 